# packages/client/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Rhizome rNet API",
    "version": "0.1.0",
    "description": "The rNet-compatible API exposed by the Rhizome store."
  },
  "servers": [{ "url": "/", "description": "Current Rhizome store" }],
  "paths": {
...
    "/rnet/v0/origins": {
      "post": {
        "operationId": "createOriginArtifact",
        "tags": ["Origin Artifacts"],
        "security": [{ "BearerAuth": [] }],
        "x-rhizome-auth": "user",
        "requestBody": {
          "required": true,
          "content": { "*/*": { "schema": { "type": "string", "format": "binary" } } }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/OriginArtifact" } }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "403": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "415": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/rnet/v0/origins/{id}": {
      "get": {
        "operationId": "getOriginArtifact",
        "tags": ["Origin Artifacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/OriginArtifact" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/rnet/v0/origins/{id}/bytes": {
      "get": {
        "operationId": "getOriginArtifactBytes",
        "tags": ["Origin Artifacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary content",
            "content": { "*/*": { "schema": { "type": "string", "format": "binary" } } }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "BearerAuth": { "type": "http", "scheme": "bearer" } },
    "schemas": {
...
      "OriginArtifact": {
        "title": "OriginArtifact",
        "description": "An immutable, UUIDv7-identified provenance record around a content-addressed payload. Ontologically inert: not media, has no kind, never appears in a Vibe's objects, never consumed by a model as content. Exists purely as ground truth so ingestion can always be re-run against original bytes.",
        "type": "object",
        "required": ["rnet_schema", "uri", "owner", "content_hash", "mime", "bytes"],
        "properties": {
          "rnet_schema": {
            "const": "0.1",
            "description": "Protocol version this origin record was written under."
          },
          "uri": {
            "type": "string",
            "description": "Immutable UUIDv7 provenance-record identity, independent of payload identity.",
            "pattern": "^rnet://origin/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "owner": {
            "type": "string",
            "description": "Immutable owner identity assigned by the store at creation.",
            "pattern": "^rnet://id/[A-Za-z0-9._~-]+$"
          },
          "content_hash": {
            "type": "string",
            "description": "SHA-256 identity of the payload bytes; not record identity or authority.",
            "pattern": "^sha256:[a-f0-9]{64}$"
          },
          "mime": { "type": "string", "pattern": "^[a-z]+/[a-zA-Z0-9][a-zA-Z0-9!#$&^_.+-]*$" },
          "bytes": {
            "type": "string",
            "description": "Retrievable payload location. Returned bytes must hash to content_hash.",
            "format": "uri"
          },
          "byte_size": { "type": "integer", "minimum": 0 },
          "label": {
            "type": "string",
            "description": "Human-readable name, e.g. the original filename.",
            "maxLength": 512
          },
          "uploaded_at": { "type": "string", "format": "date-time" }
        },
        "patternProperties": { "^x-[a-z0-9-]+$": { "description": "Namespaced extensions." } },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "code"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string", "minLength": 1 },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string" },
          "code": { "type": "string", "minLength": 1 }
        },
        "additionalProperties": true
      }
    }
  }
}